services/dispatcher/requirements.txt
redis==4.6.0
psycopg2-binary==2.9.9
requests==2.31.0
aiofiles==23.2.1
